requirements.txt
pydantic==2.11.7
pymongo==4.13.0
langchain-ollama==0.3.5
voyageai==0.2.4
google-generativeai==0.8.5
crawl4ai==0.7.1
ollama==0.5.1
playwright==1.53.0
bson==0.5.10
python-dotenv==1.1.1
typing-extensions=4.12.2